requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
mssql-django>=1.3,<1.4
drf-spectacular>=0.22.1,<0.23
pillow>=9.1.0,<9.2
uwsgi>=2.0.20,<2.1
django-cors-headers>=4.3.1,<4.4
drf-yasg==1.21.7
